requirements.txt
fastapi
uvicorn
twilio
python-dotenv
python-multipart
openai
